Turn Order and Combat Action (#15)
* TMPro package

* Ability const fix

* TurnManager renamed to CombatManager

* Tiles Filtering

* Update GridTileController.cs

* CombatAbilitiesUIManager

* Update Character.cs

* Update CombatManager.cs

* UI prefabs for Combat Abilities display

* Update CombatManager.cs

* Update CombatAbilityUIManager.cs

* Update BattleTestMap.unity

* Starting combat will enable and clear Combat Ability UI

* Sample import for Shader Graph

* Shader and Material for tiles

* Tile prefab update

* Highlight logic

* Handlers EffectTyp fix

* OnSelect/Deselect events and ability button creation

* Ability finish logic

* TryToEndTurn and RemoveActionPoint logic

* CombatAbilitiesPanel prefab

* Scene update

* Reduced the default amount of ability points to 2

* Ability: Kill -> Pistol Shot

* Ability: Move update

Move can now target adjacent tiles only

* Initialize Combat and Grid on startup

* Ability name wrapping

* Enemy prefab update

* Move Handler

* Debug log for abilities and effects

* Modified GridEffectHandler logic

Now, instead of Character-based effect distribution, GFH will try to execute abilities on each tile and their respective occupants. This is to allow DoEffect to perform effect even when no occupant are targeted (for abilities such as Move)

* DoEffect handlers update

* Additional Assert in ExecuteAbility

* Safety check for Damage handler

* Update MoveHandler.cs

* ClearTileHighlight fix

* Update DamageKineticHandler.cs

* Expose CurrentHealth from Character

* Log spawned enemies

* Log damage dealt

* Increase PistolShot damage to 3

* Prefabs update

* Receiving damage and dying logic

* RemoveActionPoint safety validation

Removing action poins should be done after performing an action or after confirming it, so logic related to verifying if a Combatant has enough AP to perform an Action, should be placed there.

This is just a safety check, to ensure that logic (when implemented) will work fine

* Death logic revamped

Added a wrapper to logic responsible for checking Death requirements, in case we would need to bypass them or branch logic based on its outcome.

Death logic now unregisters dying Combatant from a Tile it occupies.

Removed the FUCKASS comment.